requirements.txt
certifi==2018.10.15
chardet==3.0.4
click==6.7
coreapi==2.3.3
coreschema==0.0.4
coverage==4.5.1
Django==2.1.2
django-extensions==2.1.3
djangorestframework==3.9.0
drf-yasg==1.11.0
flex==6.13.2
idna==2.7
inflection==0.3.1
itypes==1.1.0
Jinja2==2.10
jsonpointer==1.14
jsonschema==2.6.0
MarkupSafe==1.0
openapi-codec==1.3.2
PyJWT==1.6.4
pytz==2018.7
PyYAML==3.13
requests==2.20.0
rfc3987==1.3.8
ruamel.yaml==0.15.76
simplejson==3.16.0
six==1.11.0
strict-rfc3339==0.7
swagger-spec-validator==2.4.1
uritemplate==3.0.0
urllib3==1.24.1
validate-email==1.3
